tests/wpt/tests/css/css-scroll-snap/scroll-initial-target/parsing/scroll-initial-target-invalid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title> CSS Scroll Snap 2 Test: scroll-initial-target-* with invalid values</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-initial-target">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <script>
    test_invalid_value("scroll-initial-target", "invalid_keyword");
    test_invalid_value("scroll-initial-target", "100px");
    test_invalid_value("scroll-initial-target", "none none none");
    test_invalid_value("scroll-initial-target", "invalid_keyword1 invalid_keyword2");
    test_invalid_value("scroll-initial-target", "100px 100px");
  </script>
</body>
</html>
